create account view done

+++ core/authentication/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render
 import datetime
 from django.shortcuts import render, redirect
-from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseRedirect
 from rest_framework import status
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
@@ -17,53 +17,69 @@ from django.contrib import messages
 
 
 
-@api_view(['GET'])
 def alert(request):
     return render(request, 'alert.html')
 
-@api_view(['POST'])
-def request_otp(request):
-    phone = request.data.get('phone')
-    
-    if phone:
-        if User.objects.filter(phone = phone).exists():
-            return Response("Phone already exists!", status=400)
-        return send_otp(phone)    
-    
-    else:
-        return Response("data_missing", status=400)
-
-
-@api_view(['POST'])
-def verify_otp(request):
-    phone = request.data.get('phone')
-    otp = request.data.get('otp')
-    
-    otp_obj = get_object_or_404(OTP, phone = phone, verified = False)
-    
-    if otp_obj.validity.replace(tzinfo=None) > datetime.datetime.utcnow():
-        if otp_obj.otp == int(otp):
-            otp_obj.verified = True 
-            otp_obj.save()
-            return Response("otp_verified_successfully!")
-        return Response("otp_not_matched!", status=400)
+def requestOtp(request):
+    if request.method == "POST":
+        phone = request.POST.get('phone')
+        
+        if phone:
+            if User.objects.filter(phone = phone).exists():
+                messages.error(request, "Phone already exists!")
+                return HttpResponseRedirect(request.path_info)
             
-    else:
-        return Response('otp_expired', status=400)
+            response = send_otp(phone)
+            messages.success(request, response)
+            return redirect('verifyOtp', phone) 
+        
+        else:
+            messages.error(request, "Phone number required!")
+            return HttpResponseRedirect(request.path_info)
+    return render(request, 'accounts/requestOtp.html')
+
+
+def verifyOtp(request, phone):
+    if request.method == "POST":
+        otp = request.POST.get('otp')
+        
+        otp_obj = get_object_or_404(OTP, phone = phone, verified = False)
+        
+        if otp_obj.validity.replace(tzinfo=None) > datetime.datetime.utcnow():
+            if otp_obj.otp == int(otp):
+                otp_obj.verified = True 
+                otp_obj.save()
+                messages.success(request, 'OTP Verified Successfully!')
+                return redirect('createAccount', phone)
+                    
+            messages.error(request, "Please Enter the currect OTP!")
+            return HttpResponseRedirect(request.path_info)
+                
+        else:
+            messages.info(request, "OTP Expired!")
+            return HttpResponseRedirect(request.path_info)
+    return render(request, 'accounts/verifyOtp.html', {"phone":phone})
     
     
-@api_view(['POST'])
-def create_account(request):
-    phone = request.data.get('phone')
-    email = request.data.get('email')
-    fullname = request.data.get('fullname')
-    password = request.data.get('password')
+
+def createAccount(request, phone):
+    if request.method == "POST":
+        fullname = request.POST.get('fullname')
+        email = request.POST.get('email')
+        password1 = request.POST.get('password1')
+        password2 = request.POST.get('password2')
+        
+        if password1 != password2:
+            messages.error(request, "Passowrd and confirm password must be same!")
+            return HttpResponseRedirect(request.path_info)
+        otp_obj = get_object_or_404(OTP, phone = phone, verified = True)
+        otp_obj.delete()
+        
+        User.objects.create(phone = phone, email = email, fullname = fullname, password = make_password(password1))
+        messages.success(request, "Account created successfully!")
+        return HttpResponseRedirect(request.path_info)
     
-    otp_obj = get_object_or_404(OTP, phone = phone, verified = True)
-    otp_obj.delete()
-    
-    User.objects.create(phone = phone, email = email, fullname = fullname, password = make_password(password))
-    return Response("account_created_successfully!")
+    return render(request, 'accounts/createAccount.html', {"phone":phone})
     
 
 @api_view(['POST'])
@@ -151,6 +167,150 @@ def reset_pass_confirm(request):
     else:
         messages.error(request, 'Linked Expired')
         return redirect('alert')
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+# @api_view(['POST'])
+# def request_otp(request):
+#     phone = request.data.get('phone')
+    
+#     if phone:
+#         if User.objects.filter(phone = phone).exists():
+#             return Response("Phone already exists!", status=400)
+#         return send_otp(phone)    
+    
+#     else:
+#         return Response("data_missing", status=400)
+
+
+# @api_view(['POST'])
+# def verify_otp(request):
+#     phone = request.data.get('phone')
+#     otp = request.data.get('otp')
+    
+#     otp_obj = get_object_or_404(OTP, phone = phone, verified = False)
+    
+#     if otp_obj.validity.replace(tzinfo=None) > datetime.datetime.utcnow():
+#         if otp_obj.otp == int(otp):
+#             otp_obj.verified = True 
+#             otp_obj.save()
+#             return Response("otp_verified_successfully!")
+#         return Response("otp_not_matched!", status=400)
+            
+#     else:
+#         return Response('otp_expired', status=400)
+    
+    
+# @api_view(['POST'])
+# def create_account(request):
+#     phone = request.data.get('phone')
+#     email = request.data.get('email')
+#     fullname = request.data.get('fullname')
+#     password = request.data.get('password')
+    
+#     otp_obj = get_object_or_404(OTP, phone = phone, verified = True)
+#     otp_obj.delete()
+    
+#     User.objects.create(phone = phone, email = email, fullname = fullname, password = make_password(password))
+#     return Response("account_created_successfully!")
+    
+
+# @api_view(['POST'])
+# def login_user(request):
+#     phone = request.data.get('phone')
+#     password = request.data.get('password')
+#     email = request.data.get('email')
+    
+#     if email:
+#         user = get_object_or_404(User, email = email)
+        
+#     elif phone:
+#         user = get_object_or_404(User, phone = phone)
+#     else:
+#         return Response("data_missing", 400)
+    
+#     if check_password(password, user.password):
+#         return token_response(user)
+#     else:
+#         return Response("incorrect_password", 400)
+
+    
+
+# @api_view(['POST'])
+# def pass_reset_email(request):
+#     phone  = request.data.get('phone')
+#     email  = request.data.get('email')
+    
+#     if phone:
+#         user = get_object_or_404(User, phone = phone)
+#         return send_pass_reset_email(user)
+#     elif email:
+#         user = get_object_or_404(User, email = email)
+#         return send_pass_reset_email(user)
+#     else:
+#         return Response("data missing!", 400)
+    
+    
+# @api_view(['GET'])
+# def reset_pass_form(request, email, token):
+#     token_instance = PassResetToken.objects.filter(user__email = email, token = token).first()
+    
+#     if token_instance:
+#         if datetime.datetime.utcnow() < token_instance.validity.replace(tzinfo=None):
+#             context = {
+#                 'email':email,
+#                 'token':token,
+#                 # 'base_url':TEMPLATES_BASE_URL,
+#             }
+#             return render(request, 'accounts/new-pass-form.html', context)
+#         else:
+#             token_instance.delete()
+#             messages.error(request, 'Linked Expired')
+#             return redirect('alert')
+#     else:
+#         messages.info(request, "Instance Not found!")
+#         return redirect('alert')
+    
+# @api_view(['POST'])
+# def reset_pass_confirm(request):
+#     password1 = request.data.get('password1')
+#     password2 = request.data.get('password2')
+#     email = request.data.get('email')
+#     token = request.data.get('token')
+#     token_instance = PassResetToken.objects.filter(user__email = email, token = token).first()
+    
+#     if token_instance:
+#         if datetime.datetime.utcnow() < token_instance.validity.replace(tzinfo=None):
+#             if password1 == password2:
+#                 user = token_instance.user 
+#                 user.password = make_password(password1)
+#                 user.save()
+#                 token_instance.delete()
+#                 Token.objects.filter(user = user).delete()
+#                 messages.success(request, "Password Updated successfully!")
+#                 return redirect('alert')
+#             else:
+#                 context = {
+#                 'email':email,
+#                 'token':token,
+#                 'error': "Passowrd and confirm password must be same!",
+#                 }
+#             return render(request, 'accounts/new-pass-form.html', context)
+    
+#     else:
+#         messages.error(request, 'Linked Expired')
+#         return redirect('alert')
     
 @api_view(['GET'])
 @permission_classes([IsAuthenticatedUser])
